Refactor app_bsky::feed
