--- atprotolib-rs/src/types/app_bsky/feed/get_posts.rs
@@ -1,20 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use super::defs::PostView;
-
-/*
-    app.bsky.feed.getPosts
-*/
-
-/*    Type: response
-    Id: app.bsky.feed.getPosts#response
-    Kind: object
-
-    Properties:
-    - posts: app.bsky.feed.defs#postView[] (JsonProperty: posts) [Required]
-*/
-#[derive(Serialize, Deserialize, Debug)]
-pub struct GetPostsResponse {
-    #[serde(rename = "posts")]
-    pub posts: Vec<PostView>
-}
