--- atprotolib-rs/src/types/app_bsky/feed/postgate.rs
@@ -1,14 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-/*
-    app.bsky.feed.postgate
-*/
-
-/*    Type: disableRule
-    Id: app.bsky.feed.postgate#disableRule
-    Kind: object
-
-    Properties:
-*/
-#[derive(Serialize, Deserialize, Debug)]
-pub struct DisableRule {}
